feat: rename `Are(TItem)` to `AreEqualTo(TItem)`
