+++ Source/aweXpect/That/Collections/ThatAsyncEnumerable.Elements.Are.cs
@@ -10,49 +10,8 @@ namespace aweXpect;
 
 public static partial class ThatAsyncEnumerable
 {
-	public partial class Elements
-	{
-		/// <summary>
-		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
-		/// </summary>
-		public StringEqualityResult<IAsyncEnumerable<string?>, IThat<IAsyncEnumerable<string?>?>> Are(
-			string? expected)
-		{
-			StringEqualityOptions options = new();
-			return new StringEqualityResult<IAsyncEnumerable<string?>, IThat<IAsyncEnumerable<string?>?>>(
-				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
-					=> new CollectionConstraint<string?>(
-						it,
-						_quantifier,
-						() => $"equal to {Formatter.Format(expected)}",
-						a => options.AreConsideredEqual(a, expected),
-						"were")),
-				_subject,
-				options);
-		}
-	}
-
 	public partial class Elements<TItem>
 	{
-		/// <summary>
-		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
-		/// </summary>
-		public ObjectEqualityResult<IAsyncEnumerable<TItem>, IThat<IAsyncEnumerable<TItem>?>>
-			Are(TItem expected)
-		{
-			ObjectEqualityOptions options = new();
-			return new ObjectEqualityResult<IAsyncEnumerable<TItem>, IThat<IAsyncEnumerable<TItem>?>>(
-				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
-					=> new CollectionConstraint<TItem>(
-						it,
-						_quantifier,
-						() => $"equal to {Formatter.Format(expected)}",
-						a => options.AreConsideredEqual(a, expected),
-						"were")),
-				_subject,
-				options);
-		}
-
 		/// <summary>
 		///     Verifies that the items in the collection satisfy the <paramref name="expectations" />.
 		/// </summary>
