+++ Source/aweXpect/That/Collections/ThatAsyncEnumerable.Elements.AreEqualTo.cs
@@ -1,0 +1,56 @@
+﻿#if NET8_0_OR_GREATER
+using System.Collections.Generic;
+using aweXpect.Core;
+using aweXpect.Helpers;
+using aweXpect.Options;
+using aweXpect.Results;
+
+namespace aweXpect;
+
+public static partial class ThatAsyncEnumerable
+{
+	public partial class Elements
+	{
+		/// <summary>
+		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
+		/// </summary>
+		public StringEqualityResult<IAsyncEnumerable<string?>, IThat<IAsyncEnumerable<string?>?>> AreEqualTo(
+			string? expected)
+		{
+			StringEqualityOptions options = new();
+			return new StringEqualityResult<IAsyncEnumerable<string?>, IThat<IAsyncEnumerable<string?>?>>(
+				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
+					=> new CollectionConstraint<string?>(
+						it,
+						_quantifier,
+						() => $"equal to {Formatter.Format(expected)}",
+						a => options.AreConsideredEqual(a, expected),
+						"were")),
+				_subject,
+				options);
+		}
+	}
+
+	public partial class Elements<TItem>
+	{
+		/// <summary>
+		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
+		/// </summary>
+		public ObjectEqualityResult<IAsyncEnumerable<TItem>, IThat<IAsyncEnumerable<TItem>?>>
+			AreEqualTo(TItem expected)
+		{
+			ObjectEqualityOptions options = new();
+			return new ObjectEqualityResult<IAsyncEnumerable<TItem>, IThat<IAsyncEnumerable<TItem>?>>(
+				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
+					=> new CollectionConstraint<TItem>(
+						it,
+						_quantifier,
+						() => $"equal to {Formatter.Format(expected)}",
+						a => options.AreConsideredEqual(a, expected),
+						"were")),
+				_subject,
+				options);
+		}
+	}
+}
+#endif
