+++ Source/aweXpect/That/Collections/ThatEnumerable.Elements.Are.cs
@@ -9,49 +9,8 @@ namespace aweXpect;
 
 public static partial class ThatEnumerable
 {
-	public partial class Elements
-	{
-		/// <summary>
-		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
-		/// </summary>
-		public StringEqualityResult<IEnumerable<string?>, IThat<IEnumerable<string?>?>> Are(
-			string? expected)
-		{
-			StringEqualityOptions options = new();
-			return new StringEqualityResult<IEnumerable<string?>, IThat<IEnumerable<string?>?>>(
-				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
-					=> new CollectionConstraint<string?>(
-						it,
-						_quantifier,
-						() => $"equal to {Formatter.Format(expected)}{options}",
-						a => options.AreConsideredEqual(a, expected),
-						"were")),
-				_subject,
-				options);
-		}
-	}
-
 	public partial class Elements<TItem>
 	{
-		/// <summary>
-		///     Verifies that the items in the collection are equal to the <paramref name="expected" /> value.
-		/// </summary>
-		public ObjectEqualityResult<IEnumerable<TItem>, IThat<IEnumerable<TItem>?>>
-			Are(TItem expected)
-		{
-			ObjectEqualityOptions options = new();
-			return new ObjectEqualityResult<IEnumerable<TItem>, IThat<IEnumerable<TItem>?>>(
-				_subject.ThatIs().ExpectationBuilder.AddConstraint(it
-					=> new CollectionConstraint<TItem>(
-						it,
-						_quantifier,
-						() => $"equal to {Formatter.Format(expected)}{options}",
-						a => options.AreConsideredEqual(a, expected),
-						"were")),
-				_subject,
-				options);
-		}
-
 		/// <summary>
 		///     Verifies that the items in the collection satisfy the <paramref name="expectations" />.
 		/// </summary>
