+++ Tests/aweXpect.Tests/Collections/ThatAsyncEnumerable.AtLeast.Tests.cs
@@ -33,8 +33,8 @@ public sealed partial class ThatAsyncEnumerable
 				ThrowWhenIteratingTwiceAsyncEnumerable subject = new();
 
 				async Task Act()
-					=> await That(subject).AtLeast(0).Are(1)
-						.And.AtLeast(0).Are(1);
+					=> await That(subject).AtLeast(0).AreEqualTo(1)
+						.And.AtLeast(0).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -45,7 +45,7 @@ public sealed partial class ThatAsyncEnumerable
 				IAsyncEnumerable<int> subject = Factory.GetAsyncFibonacciNumbers();
 
 				async Task Act()
-					=> await That(subject).AtLeast(2).Are(1);
+					=> await That(subject).AtLeast(2).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -56,7 +56,7 @@ public sealed partial class ThatAsyncEnumerable
 				IAsyncEnumerable<int> subject = ToAsyncEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).AtLeast(3).Are(1);
+					=> await That(subject).AtLeast(3).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -67,7 +67,7 @@ public sealed partial class ThatAsyncEnumerable
 				IAsyncEnumerable<int> subject = ToAsyncEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).AtLeast(5).Are(1);
+					=> await That(subject).AtLeast(5).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -83,7 +83,7 @@ public sealed partial class ThatAsyncEnumerable
 				IAsyncEnumerable<int> subject = ToAsyncEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).AtLeast(5).Are(1);
+					=> await That(subject).AtLeast(5).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -99,7 +99,7 @@ public sealed partial class ThatAsyncEnumerable
 				IAsyncEnumerable<int>? subject = null;
 
 				async Task Act()
-					=> await That(subject).AtLeast(1).Are(0);
+					=> await That(subject).AtLeast(1).AreEqualTo(0);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
