+++ Tests/aweXpect.Tests/Collections/ThatEnumerable.AtMost.Tests.cs
@@ -36,8 +36,8 @@ public sealed partial class ThatEnumerable
 				ThrowWhenIteratingTwiceEnumerable subject = new();
 
 				async Task Act()
-					=> await That(subject).AtMost(3).Are(1)
-						.And.AtMost(3).Are(1);
+					=> await That(subject).AtMost(3).AreEqualTo(1)
+						.And.AtMost(3).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -48,7 +48,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = Factory.GetFibonacciNumbers();
 
 				async Task Act()
-					=> await That(subject).AtMost(1).Are(1);
+					=> await That(subject).AtMost(1).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -64,7 +64,7 @@ public sealed partial class ThatEnumerable
 				int[] subject = [1, 1, 1, 1, 2, 2, 3];
 
 				async Task Act()
-					=> await That(subject).AtMost(3).Are(2);
+					=> await That(subject).AtMost(3).AreEqualTo(2);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -75,7 +75,7 @@ public sealed partial class ThatEnumerable
 				int[] subject = [1, 1, 1, 1, 2, 2, 3];
 
 				async Task Act()
-					=> await That(subject).AtMost(3).Are(1);
+					=> await That(subject).AtMost(3).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -91,7 +91,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).AtMost(3).Are(2);
+					=> await That(subject).AtMost(3).AreEqualTo(2);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -102,7 +102,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).AtMost(3).Are(1);
+					=> await That(subject).AtMost(3).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -118,7 +118,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int>? subject = null;
 
 				async Task Act()
-					=> await That(subject).AtMost(1).Are(0);
+					=> await That(subject).AtMost(1).AreEqualTo(0);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
