+++ Tests/aweXpect.Tests/Collections/ThatEnumerable.Between.Tests.cs
@@ -36,8 +36,8 @@ public sealed partial class ThatEnumerable
 				ThrowWhenIteratingTwiceEnumerable subject = new();
 
 				async Task Act()
-					=> await That(subject).Between(0).And(2).Are(1)
-						.And.Between(0).And(1).Are(1);
+					=> await That(subject).Between(0).And(2).AreEqualTo(1)
+						.And.Between(0).And(1).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -48,7 +48,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = Factory.GetFibonacciNumbers();
 
 				async Task Act()
-					=> await That(subject).Between(0).And(1).Are(1);
+					=> await That(subject).Between(0).And(1).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -64,7 +64,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Between(3).And(4).Are(1);
+					=> await That(subject).Between(3).And(4).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -75,7 +75,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Between(3).And(4).Are(2);
+					=> await That(subject).Between(3).And(4).AreEqualTo(2);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -91,7 +91,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Between(1).And(3).Are(1);
+					=> await That(subject).Between(1).And(3).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -107,7 +107,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int>? subject = null;
 
 				async Task Act()
-					=> await That(subject).Between(0).And(1).Are(0);
+					=> await That(subject).Between(0).And(1).AreEqualTo(0);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
