+++ Tests/aweXpect.Tests/Collections/ThatEnumerable.Exactly.Tests.cs
@@ -37,8 +37,8 @@ public sealed partial class ThatEnumerable
 				ThrowWhenIteratingTwiceEnumerable subject = new();
 
 				async Task Act()
-					=> await That(subject).Exactly(1).Are(1)
-						.And.Exactly(1).Are(1);
+					=> await That(subject).Exactly(1).AreEqualTo(1)
+						.And.Exactly(1).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -49,7 +49,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = Factory.GetFibonacciNumbers();
 
 				async Task Act()
-					=> await That(subject).Exactly(1).Are(1);
+					=> await That(subject).Exactly(1).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -65,7 +65,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Exactly(4).Are(1);
+					=> await That(subject).Exactly(4).AreEqualTo(1);
 
 				await That(Act).DoesNotThrow();
 			}
@@ -76,7 +76,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Exactly(4).Are(2);
+					=> await That(subject).Exactly(4).AreEqualTo(2);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -92,7 +92,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int> subject = ToEnumerable([1, 1, 1, 1, 2, 2, 3]);
 
 				async Task Act()
-					=> await That(subject).Exactly(3).Are(1);
+					=> await That(subject).Exactly(3).AreEqualTo(1);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
@@ -108,7 +108,7 @@ public sealed partial class ThatEnumerable
 				IEnumerable<int>? subject = null;
 
 				async Task Act()
-					=> await That(subject).Exactly(1).Are(0);
+					=> await That(subject).Exactly(1).AreEqualTo(0);
 
 				await That(Act).Throws<XunitException>()
 					.WithMessage("""
